fix/Improved UX for download for different viewing mode and user types
